+++ libft/libft.h
@@ -59,7 +59,7 @@ char			*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 char			*ft_strtrim(char const *s1, char const *set);
 char			*ft_substr(char const *s, unsigned int start, size_t len);
 
-int				ft_nbdigits_base(unsigned int nb, int base);
+int				ft_nbdigits_base(unsigned long long int nb, int base);
 
 void			ft_putchar_fd(char c, int fd);
 void			ft_putendl_fd(char *s, int fd);
